Add resource based authorization

+++ Conferences.Application/Conferences/Commands/DeleteConference/DeleteConferenceHandler.cs
@@ -1,5 +1,7 @@
-﻿using Conferences.Domain.Entities;
+﻿using Conferences.Domain.Constants;
+using Conferences.Domain.Entities;
 using Conferences.Domain.Exceptions;
+using Conferences.Domain.Interfaces;
 using Conferences.Domain.Repositories;
 using MediatR;
 using Microsoft.Extensions.Logging;
@@ -8,7 +10,9 @@ using Microsoft.Extensions.Logging;
 namespace Conferences.Application.Conferences.Commands.DeleteConference
 {
     public class DeleteConferenceHandler(ILogger<DeleteConferenceHandler> logger,
-        IConferencesRepository conferencesRepository) : IRequestHandler<DeleteConferenceCommand>
+        IConferencesRepository conferencesRepository,
+        IConferenceAuthorizationService conferenceAuthorizationService) 
+        : IRequestHandler<DeleteConferenceCommand>
     {
         public async Task Handle(DeleteConferenceCommand request, CancellationToken cancellationToken)
         {
@@ -16,6 +20,9 @@ namespace Conferences.Application.Conferences.Commands.DeleteConference
 
             var conference = await conferencesRepository.GetByIdAsync(request.Id)
                 ?? throw new NotFoundException(nameof(Conference), request.Id.ToString());
+
+            if (!conferenceAuthorizationService.Authorize(conference, ResourceOperation.Delete))
+                throw new ForbidException();
 
             await conferencesRepository.DeleteAsync(conference);
         }
